+++ _examples/view/parse-template/handlebars/main.go
@@ -9,6 +9,7 @@ func main() {
 			return "Hello, " + name + "!"
 		},
 	})
+	e.Reload(true)
 
 	app := iris.New()
 	app.RegisterView(e)