+++ _examples/view/parse-template/jet/main.go
@@ -17,6 +17,7 @@ func main() {
 	if err != nil {
 		panic(err)
 	}
+	e.Reload(true)
 
 	app := iris.New()
 	app.RegisterView(e)